re organize and simplify readme a bit
