+++ .gitignore
@@ -1,2 +1,3 @@
+# auto generated folders
 node_modules
 cache
