+++ frontend/README.md
@@ -1,0 +1,12 @@
+# Welcome to our Dapp's Frontend
+
+## Install and run commands
+
+```bash
+npm install
+npm start # this spins up a local server
+```
+
+## Questions?
+
+Refer to the main [README.md](../README.md)
