+++ frontend/src/artifacts/contracts/Greeter.sol/Greeter.dbg.json
@@ -1,0 +1,4 @@
+{
+  "_format": "hh-sol-dbg-1",
+  "buildInfo": "../../build-info/bb73eb43e98e56b3cad642ea087892e0.json"
+}
