+++ frontend/src/artifacts/hardhat/console.sol/console.json
@@ -1,0 +1,10 @@
+{
+  "_format": "hh-sol-artifact-1",
+  "contractName": "console",
+  "sourceName": "hardhat/console.sol",
+  "abi": [],
+  "bytecode": "0x60566050600b82828239805160001a6073146043577f4e487b7100000000000000000000000000000000000000000000000000000000600052600060045260246000fd5b30600052607381538281f3fe73000000000000000000000000000000000000000030146080604052600080fdfea264697066735822122065e5adfad75af7d8ec645dfa4556b90aa4051c8754bd968243c099cdb797678964736f6c63430008030033",
+  "deployedBytecode": "0x73000000000000000000000000000000000000000030146080604052600080fdfea264697066735822122065e5adfad75af7d8ec645dfa4556b90aa4051c8754bd968243c099cdb797678964736f6c63430008030033",
+  "linkReferences": {},
+  "deployedLinkReferences": {}
+}
