+++ hardhat.config.js
@@ -23,7 +23,7 @@ module.exports = {
   },
   networks: {
     hardhat: {
-      chainId: 1337
-    }
-  }
+      chainId: 1337,
+    },
+  },
 };
